[refactor] 게시글 글쓰기 - custom hook 리팩
